Add schemas and routes for user verification, ticket management, and voice settings
- Introduced new schemas for statistics, tickets, users, and voice channels.
- Implemented user enrichment logic for fetching and caching Discord member data.
- Created routes for managing scheduled events in Discord, including CRUD operations and synchronization.
- Enhanced ticket management with detailed schemas for categories, channels, and settings.
- Added response models for user search and ticket statistics.
