+++ bot.py
@@ -11,11 +11,10 @@ import discord
 from discord.ext import commands
 from dotenv import load_dotenv
 
+import bot_tasks
 from config.config_loader import ConfigLoader, normalize_prefix
 from helpers.http_helper import HTTPClient
 from helpers.task_queue import start_task_workers, stop_task_workers
-from helpers.token_manager import cleanup_tokens
-from services.log_cleanup import LogCleanupService
 from utils.logging import get_logger
 from utils.tasks import spawn
 
@@ -258,9 +257,9 @@ class MyBot(commands.Bot):
         self._track_task(spawn(self.role_refresh_task()), "role_refresh")
 
         # Start cleanup tasks
-        self._track_task(spawn(self.token_cleanup_task()), "token_cleanup")
-        self._track_task(spawn(self.attempts_cleanup_task()), "attempts_cleanup")
-        self._track_task(spawn(self.log_cleanup_task()), "log_cleanup")
+        self._track_task(spawn(bot_tasks.token_cleanup_task(self)), "token_cleanup")
+        self._track_task(spawn(bot_tasks.attempts_cleanup_task(self)), "attempts_cleanup")
+        self._track_task(spawn(bot_tasks.log_cleanup_task(self)), "log_cleanup")
 
         # Register persistent views (must happen every startup for persistence to work)
         # Import here to avoid circular import issues
@@ -328,7 +327,7 @@ class MyBot(commands.Bot):
         logger.info("Bot is ready and online!")
 
         # Alert admin channel about prefix warnings if any
-        await self._alert_prefix_warnings()
+        await bot_tasks.alert_prefix_warnings(self)
 
         if self.guilds:
             await asyncio.gather(
@@ -345,57 +344,6 @@ class MyBot(commands.Bot):
             logger.info(f"Chunked members for guild '{guild.name}' ({guild.id})")
         except Exception as e:
             logger.warning(f"Could not chunk members for guild '{guild.name}': {e}")
-
-    async def _alert_prefix_warnings(self) -> None:
-        """Send admin channel alert if there were prefix normalization warnings."""
-        if not PREFIX_WARNINGS:
-            return
-
-        await asyncio.gather(
-            *(self._send_prefix_warning_for_guild(guild) for guild in self.guilds)
-        )
-
-    async def _send_prefix_warning_for_guild(self, guild: discord.Guild) -> None:
-        """Send a prefix warning alert for a single guild if configured."""
-        try:
-            # Get admin channel from config
-            bot_spam_id = await self.services.config.get_guild_setting(
-                guild.id, "channels.bot_spam_channel_id"
-            )
-            if not bot_spam_id:
-                return
-
-            channel = guild.get_channel(int(bot_spam_id))
-            if not channel or not isinstance(channel, discord.abc.Messageable):
-                return
-
-            embed = discord.Embed(
-                title="⚠️ Prefix Configuration Warning",
-                description="The command prefix configuration had issues during startup.",
-                color=discord.Color.orange(),
-            )
-            current_prefix = get_prefix()
-            warnings = get_prefix_warnings()
-
-            embed.add_field(
-                name="Warnings",
-                value="\n".join(f"• {w}" for w in warnings[:10]),
-                inline=False,
-            )
-            embed.add_field(
-                name="Current Mode",
-                value="Mention-only"
-                if commands.when_mentioned == current_prefix
-                else f"Prefixes: {current_prefix}",
-                inline=False,
-            )
-            embed.set_footer(text="Check config.yaml prefix settings")
-
-            await channel.send(embed=embed)
-            logger.info(f"Sent prefix warning alert to guild {guild.name}")
-
-        except Exception as e:
-            logger.warning(f"Failed to send prefix warning to guild {guild.name}: {e}")
 
     async def check_bot_permissions(self, guild: discord.Guild) -> None:
         """Verify required guild-level permissions and log any missing ones."""
@@ -563,83 +511,6 @@ class MyBot(commands.Bot):
                 )
         else:
             logger.info(f"Role {role_id} missing in '{guild.name}' (already reported).")
-
-    async def token_cleanup_task(self) -> None:
-        """
-        Periodically cleans up expired tokens.
-        """
-        while not self.is_closed():
-            await asyncio.sleep(300)  # Run every 5 minutes
-            cleanup_tokens()
-            logger.debug("Expired tokens cleaned up.")
-
-    async def attempts_cleanup_task(self) -> None:
-        """
-        Periodically cleans up expired rate-limiting data.
-        """
-        # Import here to avoid circular import
-        from helpers.rate_limiter import cleanup_attempts
-
-        while not self.is_closed():
-            await asyncio.sleep(300)  # Run every 5 minutes
-            # Cleanup_attempts is an async coroutine; await it to avoid "coroutine was never awaited" warnings
-            await cleanup_attempts()
-
-    async def log_cleanup_task(self) -> None:
-        """
-        Daily cleanup of old logs based on retention policies.
-
-        Runs at the configured cleanup_hour_utc time each day.
-        """
-        await self.wait_until_ready()
-
-        # Get cleanup hour from config (default to 3 AM UTC)
-        cleanup_cfg = getattr(self, "config", {}) or {}
-        cleanup_hour_utc = cleanup_cfg.get("log_retention", {}).get(
-            "cleanup_hour_utc", 3
-        )
-
-        while not self.is_closed():
-            try:
-                # Calculate seconds until next cleanup time
-                from datetime import UTC, datetime, timedelta
-
-                now = datetime.now(UTC)
-                target_time = now.replace(
-                    hour=cleanup_hour_utc, minute=0, second=0, microsecond=0
-                )
-
-                # If target time has passed today, schedule for tomorrow
-                if now >= target_time:
-                    target_time += timedelta(days=1)
-
-                seconds_until_cleanup = (target_time - now).total_seconds()
-
-                logger.info(
-                    f"Log cleanup scheduled for {target_time.strftime('%Y-%m-%d %H:%M:%S UTC')} "
-                    f"({seconds_until_cleanup / 3600:.1f} hours from now)"
-                )
-
-                # Wait until cleanup time
-                await asyncio.sleep(seconds_until_cleanup)
-
-                # Run cleanup
-                logger.info("Starting scheduled log cleanup")
-                cleanup_service = LogCleanupService(cleanup_cfg)
-                summary = await cleanup_service.cleanup_all()
-
-                logger.info(
-                    f"Log cleanup completed: {summary}",
-                    extra={"cleanup_summary": summary},
-                )
-
-            except asyncio.CancelledError:
-                logger.info("Log cleanup task cancelled")
-                break
-            except Exception as e:
-                logger.exception("Error in log cleanup task", exc_info=e)
-                # Wait 1 hour before retrying on error
-                await asyncio.sleep(3600)
 
     @property
     def uptime(self) -> str:
