+++ web/backend/core/guild_settings.py
@@ -2,281 +2,18 @@
 
 from __future__ import annotations
 
-import ipaddress
 import json
 import logging
-import socket
 from datetime import UTC, datetime
 from typing import TYPE_CHECKING, Any
-from urllib.parse import urlparse
-
-import httpx
 
 from services.db.repository import BaseRepository
+from core.logo_validator import LogoValidationError, validate_logo_url
 
 if TYPE_CHECKING:
-    from collections.abc import Mapping
-
     from aiosqlite import Connection
 
 logger = logging.getLogger(__name__)
-
-# Constants for logo URL validation
-ALLOWED_IMAGE_EXTENSIONS = (".png", ".jpg", ".jpeg", ".gif", ".webp")
-MAX_LOGO_SIZE_BYTES = 8 * 1024 * 1024  # 8MB max
-LOGO_VALIDATION_TIMEOUT = 10.0  # seconds
-
-
-class LogoValidationError(Exception):
-    """Raised when logo URL validation fails."""
-
-    pass
-
-
-def _check_response_status(status_code: int) -> None:
-    """Check HTTP response status code and raise appropriate errors."""
-    if status_code == 404:
-        raise LogoValidationError("Image not found (404)")
-    if status_code == 403:
-        raise LogoValidationError("Access denied (403) - the image may be private")
-    if status_code >= 400:
-        raise LogoValidationError(f"Failed to fetch image (HTTP {status_code})")
-
-
-def _check_content_size(headers: Mapping[str, str], max_bytes: int) -> None:
-    """Check content-length header and raise if too large."""
-    content_length = headers.get("content-length")
-    if not content_length:
-        return
-    try:
-        size = int(content_length)
-        if size > max_bytes:
-            max_mb = max_bytes / (1024 * 1024)
-            actual_mb = size / (1024 * 1024)
-            raise LogoValidationError(
-                f"Image too large ({actual_mb:.1f}MB). Maximum size is {max_mb:.0f}MB"
-            )
-    except ValueError:
-        logger.debug(
-            "Invalid content-length header '%s'",
-            content_length,
-            exc_info=True,
-        )
-
-
-def _is_private_ip(hostname: str) -> bool:
-    """Check if hostname resolves to a private or internal IP address.
-
-    Returns True if the hostname is localhost, a private IP, link-local,
-    loopback, or cloud metadata endpoint.
-    """
-    hostname_lower = hostname.lower()
-
-    # Check for localhost variants and internal domain suffixes
-    if hostname_lower in ("localhost", "localhost.localdomain"):
-        return True
-
-    # Block obvious internal hostnames (.local, .lan, .internal)
-    if hostname_lower.endswith((".local", ".lan", ".internal")):
-        return True
-
-    try:
-        # Try to parse as IP address directly
-        ip = ipaddress.ip_address(hostname)
-        if (
-            ip.is_private
-            or ip.is_loopback
-            or ip.is_link_local
-            or ip.is_reserved
-            or ip.is_multicast
-        ):
-            return True
-        # Explicitly block cloud metadata endpoints (AWS/GCP/Azure)
-        if ip.version == 4 and str(ip).startswith("169.254.169."):
-            return True
-        return False
-    except ValueError:
-        logger.debug("Provided hostname is not a direct IP", exc_info=True)
-
-    # Resolve hostname to IP addresses
-    try:
-        # Get all IP addresses for the hostname
-        addr_info = socket.getaddrinfo(hostname_lower, None)
-        for info in addr_info:
-            ip_str: str = str(info[4][0])  # Explicitly cast to string for type safety
-            try:
-                ip = ipaddress.ip_address(ip_str)
-                if (
-                    ip.is_private
-                    or ip.is_loopback
-                    or ip.is_link_local
-                    or ip.is_reserved
-                    or ip.is_multicast
-                ):
-                    return True
-                # Explicitly block cloud metadata endpoints (AWS/GCP/Azure)
-                if ip.version == 4 and ip_str.startswith("169.254.169."):
-                    return True
-            except ValueError:
-                continue
-    except (socket.gaierror, socket.herror, OSError):
-        # DNS resolution failed - treat as potentially dangerous
-        return True
-
-    return False
-
-
-async def validate_logo_url(url: str | None) -> str | None:  # noqa: PLR0912, PLR0915
-    """Validate a logo URL is reachable and returns an acceptable image.
-
-    Args:
-        url: The URL to validate, or None to clear the logo.
-
-    Returns:
-        The validated URL (normalized) or None if clearing.
-
-    Raises:
-        LogoValidationError: If validation fails with a user-friendly message.
-    """
-    if not url or not url.strip():
-        return None
-
-    url = url.strip()
-
-    # Parse and validate URL structure
-    try:
-        parsed = urlparse(url)
-    except Exception as exc:
-        raise LogoValidationError(f"Invalid URL format: {exc}") from exc
-
-    if parsed.scheme not in ("http", "https"):
-        raise LogoValidationError("URL must use http or https protocol")
-
-    if not parsed.netloc:
-        raise LogoValidationError("URL must include a domain")
-
-    # SECURITY: Prevent SSRF attacks by blocking private/internal IP addresses
-    hostname = parsed.hostname or parsed.netloc
-    if _is_private_ip(hostname):
-        raise LogoValidationError(
-            "Cannot use private, local, or internal network addresses"
-        )
-
-    # Check file extension (case-insensitive)
-    path_lower = parsed.path.lower()
-    has_valid_extension = any(
-        path_lower.endswith(ext) for ext in ALLOWED_IMAGE_EXTENSIONS
-    )
-
-    # SECURITY: Reconstruct URL from validated components to break taint chain for CodeQL
-    # At this point we've validated: scheme is http/https, hostname is public
-    sanitized_url = f"{parsed.scheme}://{parsed.netloc}{parsed.path}"
-    if parsed.query:
-        sanitized_url += f"?{parsed.query}"
-    if parsed.fragment:
-        sanitized_url += f"#{parsed.fragment}"
-
-    # Perform HEAD request to validate reachability and content
-    try:
-        async with httpx.AsyncClient(
-            timeout=LOGO_VALIDATION_TIMEOUT, verify=True
-        ) as client:
-            response = await client.head(sanitized_url, follow_redirects=True)
-
-            # Some servers don't support HEAD, use streaming GET to avoid DoS
-            if response.status_code == 405:
-                # Use streaming GET with Range header to check headers without downloading body
-                async with client.stream(
-                    "GET",
-                    sanitized_url,
-                    headers={"Range": "bytes=0-0"},
-                    follow_redirects=True,
-                ) as stream_response:
-                    # Only inspect headers, don't read the stream body
-                    # This prevents DoS from servers ignoring Range header
-                    response = stream_response
-                    _check_response_status(response.status_code)
-
-                    # SECURITY: Re-validate the final URL after redirects to prevent SSRF bypass
-                    final_url = str(response.url)
-                    final_parsed = urlparse(final_url)
-                    final_hostname = final_parsed.hostname or final_parsed.netloc
-                    if _is_private_ip(final_hostname):
-                        raise LogoValidationError(
-                            "Cannot use private, local, or internal network addresses"
-                        )
-
-                    # Check content type
-                    content_type = response.headers.get("content-type", "").lower()
-                    # Split on semicolon to handle charset parameters (e.g., "image/png; charset=utf-8")
-                    media_type = content_type.split(";")[0].strip()
-                    valid_content_types = (
-                        "image/png",
-                        "image/jpeg",
-                        "image/gif",
-                        "image/webp",
-                    )
-
-                    if media_type not in valid_content_types:
-                        if not has_valid_extension:
-                            raise LogoValidationError(
-                                f"URL does not point to a valid image. "
-                                f"Expected image type, got: {content_type or 'unknown'}"
-                            )
-                        logger.warning(
-                            "Logo URL %s has valid extension but content-type is %s",
-                            url,
-                            content_type,
-                        )
-
-                    _check_content_size(response.headers, MAX_LOGO_SIZE_BYTES)
-            else:
-                _check_response_status(response.status_code)
-
-                # SECURITY: Re-validate the final URL after redirects to prevent SSRF bypass
-                final_url = str(response.url)
-                final_parsed = urlparse(final_url)
-                final_hostname = final_parsed.hostname or final_parsed.netloc
-                if _is_private_ip(final_hostname):
-                    raise LogoValidationError(
-                        "Cannot use private, local, or internal network addresses"
-                    )
-
-                # Check content type
-                content_type = response.headers.get("content-type", "").lower()
-                # Split on semicolon to handle charset parameters (e.g., "image/png; charset=utf-8")
-                media_type = content_type.split(";")[0].strip()
-                valid_content_types = (
-                    "image/png",
-                    "image/jpeg",
-                    "image/gif",
-                    "image/webp",
-                )
-
-                if media_type not in valid_content_types:
-                    if not has_valid_extension:
-                        raise LogoValidationError(
-                            f"URL does not point to a valid image. "
-                            f"Expected image type, got: {content_type or 'unknown'}"
-                        )
-                    logger.warning(
-                        "Logo URL %s has valid extension but content-type is %s",
-                        url,
-                        content_type,
-                    )
-
-                _check_content_size(response.headers, MAX_LOGO_SIZE_BYTES)
-
-    except httpx.TimeoutException as exc:
-        raise LogoValidationError(
-            "Timed out while validating image URL. Please check the URL is accessible."
-        ) from exc
-    except httpx.RequestError as exc:
-        raise LogoValidationError(
-            "Failed to reach image URL. Please verify the URL is publicly accessible."
-        ) from exc
-
-    return url
 
 
 BOT_ADMINS_KEY = "roles.bot_admins"
