+++ web/backend/core/schemas/__init__.py
@@ -1,0 +1,249 @@
+"""Pydantic schemas for API request/response models."""
+
+from .auth import (
+    AuthMeResponse,
+    GuildListResponse,
+    GuildPermission,
+    GuildSummary,
+    SelectGuildRequest,
+    SelectGuildResponse,
+    UserProfile,
+)
+from .errors import ErrorDetail, ErrorResponse, ErrorsResponse, StructuredError
+from .events import (
+    EventSyncRequest,
+    EventSyncResponse,
+    ScheduledEventCreateRequest,
+    ScheduledEventResponse,
+    ScheduledEventsResponse,
+    ScheduledEventSummary,
+    ScheduledEventUpdateRequest,
+)
+from .guild import (
+    BotChannelSettings,
+    BotChannelSettingsResponse,
+    BotRoleSettings,
+    DiscordChannel,
+    DiscordRole,
+    EventModuleSettings,
+    GuildChannelsResponse,
+    GuildConfigData,
+    GuildConfigResponse,
+    GuildConfigUpdateRequest,
+    GuildInfo,
+    GuildInfoResponse,
+    GuildMember,
+    GuildMemberResponse,
+    GuildMembersResponse,
+    GuildRolesResponse,
+    LogoValidationRequest,
+    LogoValidationResponse,
+    MetricsSettings,
+    NewMemberRoleSettings,
+    OrganizationSettings,
+    OrganizationSettingsResponse,
+    OrganizationValidationRequest,
+    OrganizationValidationResponse,
+    ReadOnlyYamlConfig,
+    RoleDelegationConfig,
+    RoleDelegationConfigResponse,
+    RoleDelegationPolicy,
+    VoiceSelectableRoles,
+)
+from .health import HealthOverview, HealthResponse, SystemMetrics
+from .metrics import (
+    ActivityGroupCounts,
+    ActivityGroupCountsResponse,
+    ActivityTierCounts,
+    DashboardMetricsBundle,
+    DashboardMetricsResponse,
+    GameMetrics,
+    GameMetricsResponse,
+    GameStats,
+    GameTopPlayer,
+    LeaderboardResponse,
+    MessageLeaderboardEntry,
+    MetricsLive,
+    MetricsOverview,
+    MetricsOverviewResponse,
+    MetricsPeriod,
+    TimeSeriesPoint,
+    TimeSeriesResponse,
+    TopGamesResponse,
+    UserGameStats,
+    UserMetrics,
+    UserMetricsResponse,
+    UserTimeSeriesPoint,
+    VoiceLeaderboardEntry,
+)
+from .stats import StatsOverview, StatsResponse, StatusCounts
+from .tickets import (
+    TicketCategory,
+    TicketCategoryCreate,
+    TicketCategoryListResponse,
+    TicketCategoryUpdate,
+    TicketChannelConfig,
+    TicketChannelConfigCreate,
+    TicketChannelConfigListResponse,
+    TicketChannelConfigUpdate,
+    TicketFormConfig,
+    TicketFormConfigResponse,
+    TicketFormConfigUpdate,
+    TicketFormQuestion,
+    TicketFormResponse,
+    TicketFormResponseList,
+    TicketFormStep,
+    TicketFormValidation,
+    TicketInfo,
+    TicketListResponse,
+    TicketSettings,
+    TicketSettingsResponse,
+    TicketSettingsUpdate,
+    TicketStatsResponse,
+)
+from .users import UserSearchResponse, VerificationRecord
+from .voice import (
+    ActiveVoiceChannel,
+    ActiveVoiceChannelsResponse,
+    GuildUserSettingsGroup,
+    GuildVoiceGroup,
+    JTCChannelSettings,
+    PermissionEntry,
+    PrioritySpeakerEntry,
+    PTTSettingEntry,
+    SoundboardEntry,
+    UserJTCSettings,
+    VoiceChannelMember,
+    VoiceChannelRecord,
+    VoiceSearchResponse,
+    VoiceSettingsResetResponse,
+    VoiceUserSettingsSearchResponse,
+)
+
+__all__ = [
+    # auth
+    "AuthMeResponse",
+    "GuildListResponse",
+    "GuildPermission",
+    "GuildSummary",
+    "SelectGuildRequest",
+    "SelectGuildResponse",
+    "UserProfile",
+    # errors
+    "ErrorDetail",
+    "ErrorResponse",
+    "ErrorsResponse",
+    "StructuredError",
+    # events
+    "EventSyncRequest",
+    "EventSyncResponse",
+    "ScheduledEventCreateRequest",
+    "ScheduledEventResponse",
+    "ScheduledEventsResponse",
+    "ScheduledEventSummary",
+    "ScheduledEventUpdateRequest",
+    # guild
+    "BotChannelSettings",
+    "BotChannelSettingsResponse",
+    "BotRoleSettings",
+    "DiscordChannel",
+    "DiscordRole",
+    "EventModuleSettings",
+    "GuildChannelsResponse",
+    "GuildConfigData",
+    "GuildConfigResponse",
+    "GuildConfigUpdateRequest",
+    "GuildInfo",
+    "GuildInfoResponse",
+    "GuildMember",
+    "GuildMemberResponse",
+    "GuildMembersResponse",
+    "GuildRolesResponse",
+    "LogoValidationRequest",
+    "LogoValidationResponse",
+    "MetricsSettings",
+    "NewMemberRoleSettings",
+    "OrganizationSettings",
+    "OrganizationSettingsResponse",
+    "OrganizationValidationRequest",
+    "OrganizationValidationResponse",
+    "ReadOnlyYamlConfig",
+    "RoleDelegationConfig",
+    "RoleDelegationConfigResponse",
+    "RoleDelegationPolicy",
+    "VoiceSelectableRoles",
+    # health
+    "HealthOverview",
+    "HealthResponse",
+    "SystemMetrics",
+    # metrics
+    "ActivityGroupCounts",
+    "ActivityGroupCountsResponse",
+    "ActivityTierCounts",
+    "DashboardMetricsBundle",
+    "DashboardMetricsResponse",
+    "GameMetrics",
+    "GameMetricsResponse",
+    "GameStats",
+    "GameTopPlayer",
+    "LeaderboardResponse",
+    "MessageLeaderboardEntry",
+    "MetricsLive",
+    "MetricsOverview",
+    "MetricsOverviewResponse",
+    "MetricsPeriod",
+    "TimeSeriesPoint",
+    "TimeSeriesResponse",
+    "TopGamesResponse",
+    "UserGameStats",
+    "UserMetrics",
+    "UserMetricsResponse",
+    "UserTimeSeriesPoint",
+    "VoiceLeaderboardEntry",
+    # stats
+    "StatsOverview",
+    "StatsResponse",
+    "StatusCounts",
+    # tickets
+    "TicketCategory",
+    "TicketCategoryCreate",
+    "TicketCategoryListResponse",
+    "TicketCategoryUpdate",
+    "TicketChannelConfig",
+    "TicketChannelConfigCreate",
+    "TicketChannelConfigListResponse",
+    "TicketChannelConfigUpdate",
+    "TicketFormConfig",
+    "TicketFormConfigResponse",
+    "TicketFormConfigUpdate",
+    "TicketFormQuestion",
+    "TicketFormResponse",
+    "TicketFormResponseList",
+    "TicketFormStep",
+    "TicketFormValidation",
+    "TicketInfo",
+    "TicketListResponse",
+    "TicketSettings",
+    "TicketSettingsResponse",
+    "TicketSettingsUpdate",
+    "TicketStatsResponse",
+    # users
+    "UserSearchResponse",
+    "VerificationRecord",
+    # voice
+    "ActiveVoiceChannel",
+    "ActiveVoiceChannelsResponse",
+    "GuildUserSettingsGroup",
+    "GuildVoiceGroup",
+    "JTCChannelSettings",
+    "PermissionEntry",
+    "PrioritySpeakerEntry",
+    "PTTSettingEntry",
+    "SoundboardEntry",
+    "UserJTCSettings",
+    "VoiceChannelMember",
+    "VoiceChannelRecord",
+    "VoiceSearchResponse",
+    "VoiceSettingsResetResponse",
+    "VoiceUserSettingsSearchResponse",
+]
